whfast size_t shearing sheet not yet working

+++ rebound/integrators/whfast.py
@@ -61,8 +61,8 @@ class IntegratorWHFast(ctypes.Structure):
                 ("_p_jh", ctypes.POINTER(Particle)),
                 ("_p_temp", ctypes.POINTER(Particle)),
                 ("is_synchronized", ctypes.c_uint),
-                ("_N_allocated", ctypes.c_uint),
-                ("_N_allocated_tmp", ctypes.c_uint),
+                ("_N_allocated", ctypes.c_size_t),
+                ("_N_allocated_tmp", ctypes.c_size_t),
                 ("_timestep_warning", ctypes.c_uint),
                 ("_recalculate_coordinates_but_not_synchronized_warning", ctypes.c_uint)]
